Add Scale parameter in materials
This allows to scale textures up or down from their initial size, allowing
them to tile at intervals other than 512x512. However, rendering of scaled
textures does not use mipmapping, and therefore looks rather strange for
scale factors not close to 1. The lack of mipmapping is due to abuse of a
3D texture for a 2D texture array.